% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add_adpd_targets}
\alias{add_adpd_targets}
\title{Add Targets for Alzheimer's and Parkinson's Disease}
\usage{
add_adpd_targets(
  associations,
  adfile = "databases/SGC_OpenAD_candidategenes_2020-04-21.csv",
  pdfile = "databases/SGC_parkinsonsdisease_candidategenes_2019-12-20.csv",
  filepath = NULL,
  allfields = FALSE
)
}
\arguments{
\item{associations}{dataframe. The target-disease associations}

\item{filepath}{string. Path to files}
}
\value{
Returns the \code{associations} dataframe with AD.sgc.target and PD.sgc.target as extra columns.
}
\description{
Add the targets being persued at the SGC for Alzheimers and Parkinsons
}
