% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add_genage}
\alias{add_genage}
\title{Add GenAge data}
\usage{
add_genage(associations, human = TRUE, mouse = FALSE, filepath = NULL)
}
\arguments{
\item{associations}{dataframe. The target-disease associations}

\item{human}{logical. Import human data from GenAge, a curated database that includes the few genes directly related to ageing in humans (n=3) plus the best candidate genes from model organisms}

\item{mouse}{logical. Import mouse data from GenAge, genes related to longevity and/or ageing in mice, including the max and median lifespan change}

\item{filepath}{string. Path to files}
}
\value{
Returns the \code{associations} dataframe joined with GenAge data as extra columns.
}
\description{
Add the data from GenAge, a curated database of human ageing-related genes (\code{human}) and a database of ageing- and longevity- associated genes in mice (\code{mouse})
}
\examples{
\dontrun{add_genage(associations)}
\dontrun{add_genage(associations, human = TRUE, mouse = FALSE)}
}
