% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add_longevity}
\alias{add_longevity}
\title{Add Longevity data}
\usage{
add_longevity(
  associations,
  filepath = NULL,
  all_datatypes = FALSE,
  from_file = FALSE
)
}
\arguments{
\item{associations}{dataframe. The target-disease associations}

\item{filepath}{string. Path to files}

\item{all_datatypes}{logical. If false, only the overall association score
and the genetic association score datatype columns are returned. If true,
all association scores for all datatypes are returned.}

\item{from_file}{logical. If true, read from a file, otherwise take from
within the associations dataframe}
}
\value{
Returns the \code{associations} dataframe joined with Longevity data as extra columns.
}
\description{
Add data for targets associated with longevity, based on genetic association from Open Targets
}
\examples{
\dontrun{add_longevity(associations)}
\dontrun{add_longevity(associations, all_datatypes = TRUE)}
}
