% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-targets.R
\name{multimorbidity_count}
\alias{multimorbidity_count}
\title{Get multimorbidity count}
\usage{
multimorbidity_count(
  associations_wide,
  minscore = 0.05,
  multimorbidityscore = FALSE,
  multimorbiditymean = FALSE
)
}
\arguments{
\item{minscore}{float. The minimum genetic association score to consider to be an association}

\item{multimorbidityscore}{boolean. Whether to also calculate the multimorbidity score}

\item{multimorbiditymean}{boolean. Whether to also calculate the mean of the association scores}

\item{associations}{dataframe. The target-disease associations in wide format}
}
\value{
Returns the \code{associations} dataframe with the multimorbidity count as an extra column
}
\description{
Score targets based on the number of morbidities with a genetic association
}
\examples{
\dontrun{multimorbidity_score(associations_wide)}
}
