% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-targets.R
\name{multimorbidity_mean}
\alias{multimorbidity_mean}
\title{Get multimorbidity mean score}
\usage{
multimorbidity_mean(associations_wide)
}
\arguments{
\item{associations}{dataframe. The target-disease associations in wide format}

\item{minscore}{float. The minimum genetic association score to consider to be an association}
}
\value{
Returns the \code{associations} dataframe with the mean association score as an extra column
}
\description{
Score targets based on the scores of the multimorbidity associations
}
