% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(associations, gene_code, save = FALSE, morbidities_ignore = NULL)
}
\arguments{
\item{associations}{dataframe. The target-disease associations in wide
format, in which each row is a target}

\item{gene_code}{string. Gene code for the gene of interest.}

\item{save}{logical. If true, the plot is saved, with the default name}

\item{morbidities_ignore}{vector. Morbidities to ignore when plotting}
}
\value{
Returns a heatmap showing the association of a gene with each morbidity
}
\description{
Plots a heatmap of the associations for an individual gene
}
\examples{
\dontrun{plot_heatmap("APOE")}
}
