% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmaps.R
\name{plot_multiheatmap}
\alias{plot_multiheatmap}
\title{Plot Multi-Heatmap}
\usage{
plot_multiheatmap(
  associations,
  save = FALSE,
  sort_nmorb = FALSE,
  gene_order = NULL
)
}
\arguments{
\item{associations}{dataframe. The target-disease associations in wide
format, in which each row is a target-disease association}

\item{save}{logical. If true, the plot is saved as a png}

\item{morbidities_ignore}{vector. Morbidities to ignore when plotting}
}
\value{
Returns a heatmap showing the overall association score of each
target with each morbidity
}
\description{
Plots a heatmap of the associations for a subset of genes
}
\examples{
\dontrun{plot_multiheatmap(some_associations)}
}
