% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-upset.R
\name{plot_upset}
\alias{plot_upset}
\title{UpSetR plot}
\usage{
plot_upset(
  associations = NULL,
  associations_wide = NULL,
  association_type = "genetic",
  only_longevity = TRUE,
  morbidities_intersection = list("heart_disease", "neurodegenerative_disease",
    "longevity"),
  morbidities_ignore = NULL,
  queries = NULL
)
}
\arguments{
\item{associations}{dataframe. The target-disease associations in long
format, in which each row is a target-disease association}

\item{associations_wide}{dataframe. The target-disease associations in wide
format, in which each row is a unique target and the associations with
different morbidities are columns}

\item{association_type}{string. The type of association to include in the
plot, e.g. "genetic" or "overall" associations}

\item{only_longevity}{logical. When true, only targets associated with
longevity are included in the plot}

\item{morbidities_intersection}{list. A list of morbidities for which the
intersection should be highlighted in the plot}

\item{morbidities_ignore}{vector. Morbidities to ignore when plotting}

\item{queries}{list. Custom queries}
}
\value{
Returns an UpSetR plot showing the number of targets associated with
each combination of morbidities
}
\description{
Generates an UpSetR plot for the number of targets associated with each
combination of morbidities
}
\examples{
\dontrun{plot_upset(df, morbidities_intersection = list("heart_disease","neurodegenerative_disease", "longevity"))}
\dontrun{plot_upset(associations = df, association_type = "overall", only_longevity = TRUE)}
}
