% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predigtTargetAge.R
\name{predictTargetAge}
\alias{predictTargetAge}
\title{predictTargetAge}
\usage{
predictTargetAge(book, genre, path_models, path_genre_prop)
}
\arguments{
\item{book}{Stagger tagged book}

\item{genre}{Book genre}

\item{path_models}{Path to folder where glm models are stored}

\item{path_genre_prop}{Path to folder where genre probabilities are stored}
}
\value{
Returns a predicted target age
}
\description{
A function that predicts target age of a text
}
