% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_associations}
\alias{read_associations}
\title{Read Associations}
\usage{
read_associations(
  filepath = NULL,
  association_file = paste0(filepath, "disease_associations.csv"),
  morbidity_file = paste0(filepath, "disease_list.csv"),
  controls = FALSE,
  control_association_file = paste0(filepath, "control_associations.csv"),
  control_morbidity_file = paste0(filepath, "control_list.csv"),
  remove_morbidities = NULL
)
}
\arguments{
\item{filepath}{string. Path to files}

\item{association_file}{string. The csv file containing the associations}

\item{morbidity_file}{string. The csv file containing the morbidities and EFO codes}

\item{control_association_file}{string. The csv file containing the associations for control morbidities}

\item{control_morbidity_file}{string. The csv file containing the control morbidities and EFO codes}

\item{remove_morbidities}{vector. An optional list of morbidities to remove}

\item{control}{logical. Indicates whether to also read in the control associations}
}
\value{
Returns the \code{associations} dataframe joined with morbidities
added as an extra columns.
}
\description{
Reads in the list of target-disease associations retrieved from Open Targets
for the list of morbidities.
}
\examples{
read_associations()
read_associations(controls = TRUE, controlfilename = "control_associations.csv")
read_associations(remove_morbitities = c("premature ageing", "arthritis", NA))
}
