% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_GO}
\alias{read_GO}
\title{Read in GO Process annotations data}
\usage{
read_GO(
  filepath = NULL,
  filename = "databases/QuickGO-annotations-1590524472235-20200526.tsv"
)
}
\arguments{
\item{filepath}{string. Path to files}

\item{filename}{string. Name of QuickGO annotation file}
}
\value{
Returns a dataframe containing the GO term associations
}
\description{
Read in targets associated with GO terms from QuickGO annotation file
}
