% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_longevity}
\alias{read_longevity}
\title{Read in Longevity data}
\usage{
read_longevity(filepath = NULL, all_datatypes = FALSE)
}
\arguments{
\item{filepath}{string. Path to files}

\item{all_datatypes}{logical. If false, only the overall association score
and the genetic association score datatype columns are returned. If true,
all association scores for all datatypes are returned.}
}
\value{
Returns a dataframe containing target associations with Longevity
}
\description{
Read in targets associated with longevity, based on genetic associations
from Open Targets
}
\examples{
read_longevity()
read_longevity(all_dataypes = TRUE)
}
