% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_safety}
\alias{read_safety}
\title{Read in safety data from Open Targets}
\usage{
read_safety(
  filepath = NULL,
  filename = "/databases/OT_20.04/known_target_safety-2020-04-01.json",
  from_file = FALSE
)
}
\arguments{
\item{filepath}{string. Path to files}

\item{from_file}{logical. If true, read from a file, otherwise take from
within the associations dataframe}

\item{associations}{dataframe. The target-disease associations}

\item{all_datatypes}{logical. If false, only the overall association score
and the genetic association score datatype columns are returned. If true,
all association scores for all datatypes are returned.}
}
\value{
Returns a dataframe containing target associations with Longevity
}
\description{
Read curated safety data
}
\examples{
\dontrun{read_longevity(all_dataypes = TRUE)}
}
