% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello}
\usage{
say_hello(name = "stranger")
}
\arguments{
\item{name}{string. Your name}
}
\value{
Returns a greeting, personalised using \code{name} if provided.
}
\description{
Receive a greeting from your package
}
\examples{
say_hello()
say_hello('Clare')
}
