% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-data.R
\name{widen_associations}
\alias{widen_associations}
\title{Widen Associations}
\usage{
widen_associations(associations)
}
\arguments{
\item{associations}{dataframe. The target-disease associations in long
format, in which each row is a target-disease association}
}
\value{
Returns a dataframe of the target-disease associations in wide
format, in which each row is a unique target and the associations with
different morbidities are columns
}
\description{
Reshapes association data from long format (in which each row is a target-disease
association) to wide format (in which each row is a unique target and the associations with
different morbidities are columns)
}
\examples{
\dontrun{associations_wide <- widen_associations(associations)}
}
