% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add_longevity}
\alias{add_longevity}
\title{Add Longevity data}
\usage{
add_longevity(associations, filepath = NULL, all_datatypes = FALSE)
}
\arguments{
\item{associations}{dataframe. The target-disease associations}

\item{filepath}{string. Path to files}

\item{all_datatypes}{logical. If false, only the overall association score
and the genetic association score datatype columns are returned. If true,
all association scores for all datatypes are returned.}
}
\value{
Returns the \code{associations} dataframe joined with Longevity data as extra columns.
}
\description{
Add data for targets associated with longevity, based on genetic association from Open Targets
}
\examples{
add_longevity(associations)
add_longevity(associations, all_datatypes = TRUE)
}
